+++ the_warroom/models.py
@@ -1640,6 +1640,7 @@ class Game(models.Model):
     hirelings = models.ManyToManyField(Hireling, blank=True, related_name='games')
     undrafted_faction = models.ForeignKey(Faction, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='undrafted_games')
     undrafted_vagabond = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='undrafted_games')
+    undrafted_captains = models.ManyToManyField(Vagabond, blank=True, related_name='undrafted_captain_games')
     link = models.URLField(max_length=1000, null=True, blank=True)
     video_link = models.URLField(
         max_length=500,
@@ -1836,9 +1837,11 @@ class Effort(models.Model):
     player = models.ForeignKey(Profile, on_delete=models.PROTECT, null=True, blank=True, related_name='efforts')
     faction = models.ForeignKey(Faction, on_delete=models.PROTECT, related_name='efforts', blank=True, null=True)
     vagabond = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='efforts')
-    captains = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='efforts_as_captain')
+    captains = models.ManyToManyField(Vagabond, blank=True, related_name='efforts_as_captain')
+    discarded_captain = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='discarded_efforts')
     coalition_with = models.ForeignKey(Faction, on_delete=models.PROTECT, null=True, blank=True, related_name='efforts_in_coalition')
     dominance = models.CharField(max_length=20, choices=DominanceChoices.choices, null=True, blank=True)
+    brazen_demogogue = models.BooleanField(default=False)
     win = models.BooleanField(default=False)
     score = models.IntegerField(validators=[MinValueValidator(0)], null=True, blank=True)
     game = models.ForeignKey(Game, on_delete=models.CASCADE, related_name='efforts')